refactor(validation): Extract types, rename errors.ts, add schema compilation test
